Support other LaTeX grammar providers
The original checked if the grammar `name` equalled `LaTeX`, which does not follow any sort of regulation and cannot be expected to be true for all LaTeX language providers.

The proposed change replaces the check with `scopeName`, which is far more consistent across language packages; if a grammar is for LaTeX, the scope name will be `text.tex.latex`. I also changed the equality to `===`, as the checking is between two strings and seems more appropriate.